dd-16 Separate modules into classes
