+++ app/src/main/java/io/krugosvet/dailydish/android/dagger/Component.kt
@@ -1,6 +1,9 @@
 package io.krugosvet.dailydish.android.dagger
 
 import dagger.Component
+import io.krugosvet.dailydish.android.dagger.module.AccountModule
+import io.krugosvet.dailydish.android.dagger.module.AppModule
+import io.krugosvet.dailydish.android.dagger.module.NetworkModule
 import io.krugosvet.dailydish.android.mainScreen.MealListPageFragment
 import io.krugosvet.dailydish.android.mainScreen.StartupActivity
 import io.krugosvet.dailydish.android.utils.baseUi.BaseActivity
